first automatic multi executor commit
introduce a function to compute the number of connected components of the undirected graph (boid groups, or clusters).
introduce cluster and executor classes.
every time a new group is formed, create an executor (the one that has been inactive for longer). if two executors meet on the same boid retain the younger one.

introduce z key to kill a specific boid